Large commit!! I splitted the saga and action codes into 7 groups, and modify some parts such as comments re-rendering and list page design. But I haven't yet completed splitting, since I couldn't have splitted the reducer code. It needs more discussion.
