+++ front/src/components/molecules/CommentList/index.js
@@ -4,11 +4,12 @@ import { Button, Comment, Form, Header, Modal, Loader } from 'semantic-ui-react'
 // username : 유저 아이디 (해시값을 얻기 위해 필요)
 // password : 유저 패스워드 (해시값을 얻기 위해 필요)
 // check_meeting_click : 댓글 리스트 로드 여부
+// comments : 댓글 목록 (댓글 목록을 불러올 때마다 리렌더링하기 위해 필요)
 // meeting_id : 미팅 게시물 고유값
 // add_comment_click : 댓글 작성할 때 액션을 디스패치할 함수
 // edit_comment_click : 댓글 수정할 때 액션을 디스패치할 함수
 // delete_comment_click : 댓글 삭제할 때 액션을 디스패치할 함수
-export const CommentList = ({ username, password, check_meeting_click, meeting_id,
+export const CommentList = ({ username, password, check_meeting_click, comments, meeting_id,
                               add_comment_click, edit_comment_click, delete_comment_click }) => {
   console.log('<CommentList Rendering>')
   const hash = new Buffer(`${username}:${password}`).toString('base64')  //  유저의 해시값
