+++ front/src/components/molecules/KindList/index.js
@@ -2,16 +2,36 @@ import React from 'react'
 import MeetingEntry from '../../../containers/MeetingEntry'
 import { Segment, Card, Header, Label } from 'semantic-ui-react'
 
+// 모임 유형에 따라 다른 라벨 태그 생성
+const get_label = (kind) => {
+  if (kind == 0)
+    return <Label as='a' color='yellow' ribbon> 음식배달 </Label>
+  else if (kind == 1)
+    return <Label as='a' color='orange' ribbon> 택시합승 </Label>
+  else if (kind == 2)
+    return <Label as='a' color='blue' ribbon> 공동구매 </Label>
+  else if (kind == 3)
+    return <Label as='a' color='pink' ribbon> 스터디 </Label>
+  else if (kind == 4)
+    return <Label as='a' color='green' ribbon> 운동 </Label>
+  else if (kind == 5)
+    return <Label as='a' color='red' ribbon> 미팅 </Label>
+  else
+    return <Label as='a' color='black' ribbon> 기타 </Label>
+}
+
 // meetinglist_list : List 페이지에서 보여줄 미팅 리스트 정보
 export const KindList = ({ meetinglist_list }) => {
   console.log('<KindList Rendering>')
   if (meetinglist_list != null) {
     let meetings = JSON.parse(meetinglist_list).results  //  미팅 리스트
+    const kind = window.location.pathname[6]
+    const label_tag = get_label(kind)
     return (
       <Segment raised>
-        <Label as='a' color='pink' ribbon> Now </Label>  {/* 라벨 */}
-        <Header as='h1'> 모집 중인 모임 </Header>         {/* 제목 */}
-        <Card.Group>                                     {/* 각각의 카드는 하나의 미팅 엔트리를 담고 있음 */}
+        {label_tag}                               {/* 라벨 */}
+        <Header as='h1'> 모집 중인 모임 </Header>  {/* 제목 */}
+        <Card.Group>                              {/* 각각의 카드는 하나의 미팅 엔트리를 담고 있음 */}
           {meetings.map(meeting_entry =>
             <Card key = {meeting_entry.id} >
               <MeetingEntry meeting_info = {meeting_entry}/>
