+++ front/src/containers/Page/AllPage.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux'
 import { AllPage } from '../../components/organisms/AllPage'
-import { change_page_num_action } from '../../store/snu_moyeo/actions'
+import { change_page_num_action } from '../../store/snu_moyeo/MeetingLoad/actions'
 
 const mapStateToProps = (state) => {
   return {
