+++ front/src/store/snu_moyeo/selectors.js
@@ -17,6 +17,7 @@ export const initialState = {
   meetinglist_list : sessionStorage.getItem("kind"),
   meetinglist_all : sessionStorage.getItem("all"),
   report_list : sessionStorage.getItem("report_list"),
+  comments : sessionStorage.getItem("comments"),
   page_num : sessionStorage.getItem("page_num"),
 
   check_leader_click : false,
